Refactor: Melhoria no Delete

+++ src/main/java/com/movieflix/controller/MovieController.java
@@ -10,6 +10,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/movieflix/movie")
@@ -56,7 +57,11 @@ public class MovieController {
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> delete (@PathVariable Long id){
-        movieService.delete(id);
-        return ResponseEntity.noContent().build();
+        Optional<Movie> movie = movieService.findById(id);
+        if (movie.isPresent()){
+            movieService.delete(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
